+++ src/__init__.py
@@ -2,5 +2,5 @@
 metadata for the package
 """
 
-__version__ = "1.4.4"
+__version__ = "1.4.5"
 __author__ = "Edmundo Garcia: ingega@gmail.com"